+++ src/refactoring/components/cart/CartItemCard.tsx
@@ -1,4 +1,4 @@
-import { CartItem } from "../../types";
+import { CartItem } from "../../../types";
 
 interface CartItemCardProps {
   item: CartItem;
