+++ src/refactoring/components/cart/ProductCard.tsx
@@ -1,4 +1,4 @@
-import { Product } from "../../types";
+import { Product } from "../../../types";
 
 interface ProductCardProps {
   product: Product;
